pyproject.toml
[tool.poetry]
name = "music-db-finder"
version = "0.1.0"
description = "Music db finder"
authors = ["Tomislav Nazifovic <tomislav.nazifovic@gmail.com>"]
license = "GNU"
packages = [{include = "music_db_finder", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
rapidfuzz = "^3.9.3"
spotipy = "^2.24.0"
ytmusicapi = "^1.7.3"
python3-discogs-client = "^2.7"
ipykernel = "^6.29.4"
rich = "^13.7.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-cov = "^5.0.0"
ruff = "^0.5.4"
pytest-ruff = "^0.4.1"
mypy = "^1.11.0"
pytest-mypy = "^0.10.3"
isort = "^5.13.2"
black = "^24.4.2"
flake8 = "^7.1.0"


[tool.pytest.ini_options]
addopts = "-rfxX --verbose --cov-report html --cov-report term-missing --cov=src"
#addopts = "-rfxX --mypy --ruff --verbose --cov-report html --cov-report term-missing --cov=src"
testpaths = ["tests", "src"]
markers = ["mypy"]


[tool.ruff]
exclude = [".git", ".gitlab", ".mypy_cache", ".pytest_cache", ".ruff_cache", ".vscode"]
line-length = 100
target-version = "py311"
output-format = "full"


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]


[tool.ruff.format]
docstring-code-format = false


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
